backend/requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.3.0

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Authentication (para futuro)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
httpx==0.25.2
